platform-px4: Add the PX4Plus
The question was asked and its an interesting one as the PX4Plus is probably
right on the limit of what we can handle if we use the ROM pods.

This is untested but builds and also not very useful as it doesn't yet have
the vt driver, keyboard, font or floppy. It does seem to prove that if you
breathe in then you can fit the Z80 core into 32K of ROM with 8K for the
common/udata and put the video/floppy/fonts banked into the top of the RAM
area. That would still allow something like 48K binaries to run, and as the
PX4Plus has at best 128K of RAMdisc for swapping 32K is probably a saner
limit anyway.

And no.. I do not think swapping to a 38400 baud serial floppy would rock.