pcw8256,fdc765: not the architectural question around banking and swap
Basically the problem is this:

On a 16K banked machine we have 0x0000-0xBFFF holding user code and we can
switch between processes low 48K with gay abandon. However the top 16K holds
process memory, common, *and* our stacks.

What probably needs to happen is that either we bounce such buffers (slow) or
teach the swap device driver in the platform to do bank flipping.

The problem with this last one is that
- it has to be done per platform
- some machines have brain damaged banking where a given address in a bank
  cannot be remapped to a different virtual address

Unfortunately it's not obvious how to create a generic solution that doesn't
punish sensibly designed machines.