msx2: various platform side and vdp fixes
With these and the small core fixes to follow we blow up for reasons I don't
yet understand in the uput() call in create_init. Presumably some kind of
mapping setup bug. With that commented we hit the boot device request, at
which point we have no keyboard.

Display is still wonky. Guess it wants an MSXpert to figure out whats wrong
with the 80x25 setup