bank32: make passive provision for awkward boxes with low 32K switching
This allows you to define them as a 32K paged system and use the provided
swap and page hooks, plus your switch functions to copy blocks between the
spare parts of the high 32K and the pages they are stashed in.

This means that with a typical usage model of a single large app performance
will basically be as good as a sensible layout, and only if you have two
larger apps actively running at once will it degrade much.

This will be pretty much essential to cover the uBee and the N8VEM-2.