z80fixedbank: Always use the swapstack
If we don't then when we call map_kernel after copying the udata back we
will scribble somewhere on the stack of the process. We could potentially
move the switch back to the point we've fixed up sp, but as we have a swap
stack we might as well use it.

If we get banked boxes without swap then it may be worth looking at the other
path for a lib/z80fixedbanknswap.s