execve: Fix a load of problems in both crt0.s and the kernel
The good news is this fixes a lot of things. The bad news is that all the
old binaries were broken. Oh well.. if we are going to have binary breakage
we might as well do it twice in one day.

Data, bss and break are now all set up correctly (I hope)