inode: socket hooks
On read we use the pipe logic for sockets, on write they work rather like a
device. This slightly odd set up means that we can buffer incoming packets
through the disk cache and thus onto disk if needed. It's also very cheap
from a code size point of view.