ptab_end: redo the various process table scans
Currently we do a compare with &ptab[maxproc], which generates a pile of
maths to scale maxproc each time around the loop. This is *silly* and its
repeated in various places where it takes space, time and registers.

Instead in start.c evaluate &ptab[maxproc] once after maxproc is computed
and save it as ptab_end. Our comparisons now now don't do the math. This
saves us 180 bytes, or nearly half the cost of adding flock support.