tgl6502: initial bring up tree elements for the TGL6502 platform
There is a lot left to fill in on the 6502 tree. There are also some very
weird bugs right now such as kprintf working/not working depending upon subtle
changes to code sizes elsewhere. It's hard at this point to tell if its
emulator bugs, something weird in the setup code, compiler etc.

All part of the fun.