libc: Make the libc code with cc65. Clean all the header bugs it found
cc65 is quite pedantic about consts, sdcc is rather slack. Clean up all the
errors cc65 found. That should also make it rather easier to do the gcc
builds.

This isn't yet a complete cc65 libc. It doesn't handle some of the calls due
to arg ordering funnies and it doesn't have a syscalls lib. A pure runtime
library for a cc65 'fuzix' target and crt0.s are also needed.