execve: use direct user mappings
If your disk devices don't support this then you'll need to enable the
CONFIG_LEGACY_EXEC option and fix the target. (This probably will get renamed
because if we get to stuff like 68000 it may be we have a cache big enough to
make execve via it useful)

This makes execve much faster, way faster on slow memory devices. The remaining
performance killer is the lack of vfork.