banking: Strip out all the experiments with CODE1 etc
By now it's obvious that banking needs to be done per platform and the logic
really wants to be in the linker. Therefore get rid of all the annoying noise.

In addition with the patches to cc65 we don't need __fastcall__ hacks, and
in fact as we can't fastcall a banked 6502 function we don't want fastcalls
leaking anywhere explicit.