zx128: arrange the memory map with a hole for the 5C00-5D3A area
Some ROMs (eg the betadisk interface) combine an arrangement whereby the
I/O ports are only accessible from the ROM with a set of ROM routines that
blindly assume spectrum basic is present and wish to pee in the system
variables area.

Whether this will be sufficient or hacked ROMs will be needed its hard to
tell given that we also need interrupts to stay off while in the foreign
ROM (otherwise our IM2 code will blow up).