libs: Support Z80 variant builds
Allow PLATFORM to be set and make it use a different crt0 and syscall.s file.
We need this for platforms like ZX128 where we can't use the normal RST calls.

For relocatable binaries when we get there the syscall vector is going to need
to be discoverable - perhaps passed in a register or on the passed stack ?