px4plus: Initial cut of banking support
Nothing runnable yet but we now build the PX4/PX4 Plus as a set of banked ROM
images which is what is needed. The actual box has

64K RAM (some used for video)
32K OS ROM (CP/M) - overlays low 32K if switched in
2 x up to 32K ROM - overlay 0x6000-0xDFFF

The "intelligent" RAMdisc and the sidecar RAM discs are not memory mapped but
can be used for task switching/swap.

The floppy drive is serial attached at a whopping 19200 baud.