lowlevel-6809.s: ashlsi3 workaround
When calling ashlsi3 (taken from the libgcc source) the arguments
are not set up on the stack as the ashlsi3 code expects.

It might be a bug in gcc-6809, but until we get this sorted out,
use this little prologue to have the function work correctly.

This was the last hurdle to get a relatively usable system going.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>