sh: correct end and use sbrk() rather than v7 linkages
With this change the v7 Bourne shell now runs, even on 32K userspace boxes.

Later she added quite a few things (see
http://heirloom.cvs.sourceforge.net/heirloom/heirloom-sh/) however most of
those are memory expensive.

Probably we should add the use of '#' as a comment character as this is
nowdays considered normal.

Of the other features - an external test suits our memory size better,
while the later builtins aren't that useful and again have size issues. Only
unset cannot be external and might be useful.

Of the late features it might be nice to get 8bit clean

-