swap: introduce swapping onto IDE, add the needed ZX128 bits
The concept is:

	swap_map(x) turns x into an address that will hold the data to be
written out.
	swap_page holds the "page" which is a cookie passed from the bank
manager to the swapper and on to the drivers.
	map_for_swap takes the cookie and puts the page in the required space
	map_kernel_{restore} then goes back to kernel mappings

The bank manager is responsible for not queuing a single I/O that spans two
"pages"

This then allows for swapping support, including banked memory from bank16k and
the like to use a generic shared driver without platform specifics leaking into
the driver layer code.