trs80: Add support for two consoles (shift-F1 and shift-F2)
Useful for debugging work and testing interactions.

The TRS80 cannot as far as I understand move the physical video base, so
we exchange the memory banks and do some additional pointer flipping and
the like to get it to work.

Cursor handling is still not quite right and needs debugging.