FUZIX: ssh.c: environment assignment bugfix + prettier prompting
here's a patch to ssh.c that:

1. Sets the prompt to include a space
2. Prints a newline in the output of the built-in "pwd"
3. Fixes environment variable assignment

As to no. 3:  The original code used "putenv()", which, on a new
variable, inserts the tmp char *  in to the env[] array.  "ssh"
proceeded to reuse this string for more input, thereby screwing up the
environment.  I changed it to "setenv()" which makes a new copy of the
string, then inserts it into env[].

This way at least us poor "ssh" users can set PATH until the real "sh"
and "init" are working better for the 6809.

--
Brett M. Gordon,
beretta42@gmail.com