udata: optimise the udata handling further for bankfixed
Take the idle in the thread doing switchout() and make it act simply as a
function call (no udata switch) providing

- We go from just us running to idle
- We are the only process woken

This saves us a lot. It has to be per arch and per banking type but it does
live entirely in switchout() and is invisible to the core so has no affect
on anything it doesn't work with or where it makes no sense.

Also add it for the 6809, although the 6809 changes from nready all still
need debugging first, so its disabled.