dragondos: get read side of driver working
With these changes we can now read dragondos floppies fairly reliably. We
need to optimise the wait for the initial DRQ as now and then we don't quite
make it in the clocks allowed and end up retrying

Writes blow up at the end of the second write in a FIRQ storm, not sure why
yet. Possibly its up to us to disable FIRQ in the NMI handler ?