graphics: Have a rethink
Basically replace all the operations with a single tight kernel operation which
can be used to construct them all with reasonable efficiency, and is good at
certain useful things. Push all the alignment considerations onto the user
code so that everyone doesn't have the time and size penalties for it when
they don't care.

Redo the NC100 to follow this implementation.

trs80 needs redoing yet but it will take the trs80 hires graphics from
unbearable to slow 8)