dragon-nx-32: add autoprobe of HDBDOS
We now identify the HDBDOS ROM by the signature, and then pull out the MMIO
port, SCSI LUN and identify the controller type so that we can set the
cartridge type nicely.

Do need to move most of main.c into discard as its getting a bit big, but it's
all discardable so that's fine.