dragon-nx32: add first test bits of scsi support
We now scan 0xFF7x for a TC3 controller and maybe even probe it properly but
not yet tested out.

Also sort out the discard segment properly. Right now it's not important but
at some point we'll find a way to reuse that space nicely.

Still lots to do including the proper autodetect for the TC3 and bank
switching the TC3 in/out if (as it basically has to be) it's in an MPI, and
handling swap or userspace I/O (so even running programs off it won't work
yet in the unlikely event it gets that far)