flat_mem: import the code based upon MAPUX
Why invent when you can recycle 8)

This hopefully forms the basis of the bits we need for supporting devices
with a flat memory space, while handling fork and memory compaction and other
horrors a little bit.

It's a starting point. There are lots of things that would improve the
performance, and it can't handle multiple memory ranges yet.