networking: first cut at the native TCP/IP stack interface
This is far from finished or even building yet, but now has the right sort
of form. For the user space side the current plan is to plug it into
something like uIP, LwIP or Harry Kalogirou's ktcp for ELKS.

In theory the API is such that beyond "what fits" it shouldn't matter whether
we bind a teeny/tiny stack like A J Kroll's microcontroller TCP/IP or something
a bit more refined.