alloc_socket: interrupt lock
In the normal case this is safe because we always allocate from in kernel.
Likewise it's safe for accept with the native TCP/IP. However it's not safe
if we have a hardware TCP/IP and it implements accept() and thus allocates
accepting sockets from an interrupt.

So block IRQs for the brief scan.

[noted by Brett]