start: allow for argument parsing by system and platform
Note: this re-orders a few bits of input and device_init(). Hopefully that
won't break anyone. It's needed so that we can make device_init depend upon
the parsed parameters on the command line.

We also accept ro/rw for the root mount type and we pass unknown arguments on
to init.

It's not entirely useful as we don't yet support remounting!