ue: Rework undo to be more 16-bit int friendly but only 7-bit clean
It was flagging "delete" versus "insert" by negating the int storing
the pointer. This might be troublesome for us where int is 16 bit
and the pointer might be above 0x8000.

We'll instead use bit 7 of the character in question.

Some possible alternatives are using an extra byte per undo step,
or storing a 15-bit offset from the buffer start together with the
delete/insert flag.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>