Applications/*/Makefile.6809: Tidy up dependencies in rules
Avoid having all objects depend on all source files. Instead use
better defined static rules that contain the dependencies.

By using $^ in the APPS rules the CRT0 is included because
the APPS depend on it. Makes it easy to let e.g. APPSNS depend
on CRT0NS instead.

Also fix the size.report target so that it gets built only when
needed.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>