libm: the backbone of the FP maths support
Only actually using the single precision paths in the current platforms as the
SDCC Z80 port only supports "float", so we define double as float and #define
over the methods.

Anyone coming to this with a bigger box (can we do real doubles on 6809?) will need
to do a fair chunk of further work.