scc: more tidying up
Really for 6809 it needs some work on primary() and eval() to allow for the
fact the 6809 unlike the 8080/5 can actually generate useful stack relative
accesses so the whole

	get address
	save address
	get value
	do stuff
	pop address
	store

sequence really ought to work like a global or static but generate ,s
relative addressing.

Z80 might also benefit from this and use ix relative, but it's less clear,
ditto 6801 especially as it has the tsx instruction.