net_native: lots of fixes
Merge a load of fixes from Brett Gordon into the net_native code. With these
applied it might even be usable!

There are a couple of changes from Brett's original - we want to report EOF
not EPIPE on reads of a closed socket, and we want to assume that the shutdown
flags will be set by the daemon rather than add extra closed socket checks.