marksman: first pass at a modern take on manual pages
Stuff nroff, it's slow complicated and produces wonderful typeset paper manual
pages when what we actually need are fast online ones.

manscan is a simple tool to hoik blocks of markdown out of code

marksman is a first pass at a tool for outputting markdown in a sort of man
page format.

It understands the line based mark ups

````

1. Foo
2. Foo

* Foo
* Bar

+ Foo
+ Bar

- Foo
- Bar

Along with double space indenting

[link][foo]
[link](foo)
[link]
[link]: foo

![text][picture]
![text](picture)

````

inline markup: * ** _ __ ~ ~~

| table |

but not a lot of the nicer formatting options, and it doesn't handle lengths
correctly yet either.

It's missing subtables, inline ````, inline HTML (as we don't use HTML) and
some other oddments

It also supports \* etc to quote meta symbols for inline formatting and it
has an extension for name/content (HTML DT/DD style) lists