68000: make fork a special case
By defining _fork() as trashing all registers we can implement multiple udata
pointer based 68K platforms without having to save the CPU registers on syscall
entry. Instead the child does a direct supervisor a7 fixup and rte to user space
with d0/d1 correct. The rest is handled by the wrapper.

Now this isn't perfect and we probably need to adjust the rules a bit if
anyone is doing clever register resident relative code because if you take a
signal handler just at the wrong moment.

I guess as we have to save A5, you'll have to use A5 8) for Amiga style
resident code if we ever get that far.