scc: turn arguments the right way round
Small C has an irritating design feature that means it stacks arguments left
to right not right to left. Rather than working around this use a spool file
so we can put the code fragments in right to left order.

This totally changes the ABI, which is fine because we aren't using it for
anything and the old ABI is a PITA. With this done we can actually write sane
syscall stubs.

Would benefit from optimisations to also use the internal buffer as a stack
not just flush to the spool file as we stack code fragments.

Needless to say all the old varargs gunge is now a) broken and b) not needed