Kernel: Z180 platforms (p112, n8vem-mark4): New devrd driver.
New driver supports three devices:
  Block device /dev/rd0 (minor 0) -- a RAM disk
  Block device /dev/rd1 (minor 1) -- a ROM disk
  Character device /dev/physmem (minor 2) -- access to all of physical memory

Configuration is now handled entirely within config.h, simply set
DEV_RD_ROM_PAGES and DEV_RD_RAM_PAGES to the number of 4KB pages to be
allocated to each block device. A "make clean" is required after
changing these values.

The ROM disk runs up to the end of the physical ROM, any space not used by the
ROM disk is located at the start (where the bootstrap or kernel image is
located).

/dev/physmem is a character device which allows userspace applications
to use lseek(), read() and write() to access any address in the CPU
physical memory map.

p112: Note that the ROM disk is not useful on P112 as the ROM is limited to
32KB in size and is unmapped from the CPU address space.