Kernel: zeta-v2: Simplify and improve boot code.
The kernel image is unlikely to grow larger than 48KB so we can
simplyify the process for relocating the kernel to the correct memory
banks when booting.

Now that we have ROM disk support we can drop copying the RAM disk
contents on boot (userspace could do this by reading from /dev/physmem
or /dev/rd1 if it were still desirable).

We now stop the floppy drive motors on boot so they don't spin forever
where we have booted from floppy disk.