Kernel: Add "fuzix-loader" boot mechansim
This provides a way to build a Fuzix application that loads and boots a
different Fuzix kernel version. This is very useful for testing new
kernels -- just load the file onto a running Fuzix system and run it. It
will sync the disks, steal control from the running kernel, and set the
new kernel booting.

This mechanism requires the kernel to support booting from arbitrary
locations in memory.

This is analagous to the existing "cpm-loader" mechanism and re-uses the
same tools.

Tested and confirmed working on p112, n8vem-mark4 and zeta-v2.