msx2: save map_table before switching pages
If an interrupt happens between setting the mapper register
and the update of map_table, map_save will take the wrong
input for restoring the mapping when returning from the interrupt.
By saving the map_table at the beginning we ensure that no matter
when the interrupt happens, map_restore sets the right pages.

This prevents some random hangs when doing lots of usermem operations.