65c816: first sketch of needed code
Lots to fill in yet but this is the basic idea

- Load kernel somewhere
- Bank 0 gets the CPU stacks
- Each process gets a bank
- Kernel gets stuffed in bank 0 if it'll fit, somewhere else if not
- C code runs with cc65 so it's using an external C stack in its own bank
  which should mean we don't have to worry about funny stack accesses (and
  can fix any in the support lib)

Processes aren't allowed to do long operations, or to save/restore bank
registers. Only awkward one here is the block move ops. We probably need
a blockmove 'syscall' that puts in th banks and avoids a process moving
bank while mvn/mvp executes (as it's interruptible).

As there is only one configuration of banking we can hopefully put all the glue
in the shared file to make platform support easier.