uanme: extend the sysinfo block and add swap bits to it
Yes this breaks uname() calls so one or two things need recompiling but better
to do it now than later.

For the moment it adds the CPU identifier, more space for CPU info data,
space for some expansion and also gets ready to add the swap info so free
can show swap state.