fsck: get the fsck tool working
This (with the needed kernel changes) seems to give us the correct behaviour
including handling recovering the live root file system.

If you boot with 'r' as an option or a dirty rootfs you'll now get a r/o
mount and you can fsck /dev/whatever to check it. If it's bad you'll then
get to fix it and it'll reboot.