asz80: finish conversion
This now behaves rather more like we need. It lacks includes and some other
nice to have bits but is also small enough to run on a Z80 box sanely. Trimming
out stdio and soem other slack could be done if needed but is a fair bit of
work.

Also add an nm mostly at this point to check the output is valid.

Now all we need is to tweak the smallc output syntax to match, and a linker...