netd: rework logic for closedown
Implement a poll function in uip for a user requested callback. That allows
us to rework all the polling close stuff rather more nicely and the end
result now seems to avoid getting stuck.

Also when we are asked for SS_CONNECTING move to it as we were forgetting
to ack the state change.