fsck: Fix sign when reporting blocks and inodes and some const bits
With this fixed we report valid numbers on a big fsck. There is a weird 6502
oddity that it reports "0" for the memory pool, but I think that's a cc65
bug not an fsck one.

Also fix the path allocating on the recursive search. We now don't buffer
overrun given a long path.