v65c816-big: get to the point it starts up
With these changes the image is loaded and appears to get put in the right
places. We enter the discard code hit start, hit d_open and then the compiler
generates self modifying code

	store to jmpvec+1/+2
	jmp jmpvec

So it looks like  little bit of compiler hacking will be needed next, or
possibly a dirty hack of putting jmpvec in stubs in bank 1 and bank 2 so
that it writes the address into bank 2 and jumps to a bank 1 stub that can
then fix up the mess.

(Even better of course would be if someone was nuts enough to port the ANSI
pcc compiler to 65C816)