ubee: First chunk of floppy work
This won't work yet as the core floppy asm code is still for the TRS80 which
isn't quite the same beast.

(There also appear to be some funnies with the emulator that cause problems
as it thinks the controller will provide data immediately after the command
whereas a real disk even if it were perfectly aligned would have to process
all the sector marks *before* data arrives.