msx2: Use DEVTTY for kernel messages
Commit 3616d6d added support for multiple vt in msx2,
but in the process it changed the console to the debug
port of openmsx. Since this is not a device present in
actual computers, any error in the boot process was
silently ignored. This patch reverts the console to
TTYDEV and at the same time it prints all the messages
to the debug port of openmsx.