startrek: continue clean up
There are corner cases on the torpedo tracking and where the computer solution
is wrong. The whole positioning logic is a bit messed up and I think the
tracking corner cases are due to that. Also of course Klingons are in integer
positions whereas presumably they should be at 0.5/0.5 because they are
notionally in the middle of the square.

Would it make sense to just Bresenham the movement and torpedo logic ?